+++ clang/learn-header-file/max.h
@@ -1,1 +1,1 @@
-int max(int a, int b);
+double max(double a, double b);
